requirements.txt
flask==3.0.3
flask-cors==4.0.0
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4

structlog==24.1.0
loguru==0.7.2
prometheus-client==0.20.0
python-dotenv==1.0.1

transformers==4.43.3
torch==2.3.1
accelerate==0.33.0
bitsandbytes==0.43.1
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.1
redis==5.0.7

pdfminer.six==20231228
python-docx==1.1.2
langdetect==1.0.9

requests==2.32.3
sseclient-py==1.8.0

pytest==8.3.2
flake8==7.1.0
black==24.8.0
isort==5.13.2
